[Y26W2-232] refactor(web): scroll 이벤트 커스텀 훅 분리
